fixed satuation

+++ src/main/java/baguchan/frostrealm/capability/FrostLivingCapability.java
@@ -287,7 +287,7 @@ public class FrostLivingCapability implements INBTSerializable<CompoundTag> {
 
 	public void deserializeNBT(HolderLookup.Provider provider, CompoundTag nbt) {
 		this.temperature = nbt.getInt("Temperature");
-		this.temperatureSaturation = nbt.getInt("TemperatureSaturation");
-		this.exhaustionLevel = nbt.getInt("TemperatureExhaustion");
+		this.temperatureSaturation = nbt.getFloat("TemperatureSaturation");
+		this.exhaustionLevel = nbt.getFloat("TemperatureExhaustion");
 	}
 }
